[refactor]: 로그인 처리로직을 한곳으로
