+++ frontend/src/utils/auth.js
@@ -1,33 +1,6 @@
-import Cookies from 'js-cookie';
+import Toast from '../components/Toast/Toast';
 import axiosInstance from './api';
-import { state, setState } from './store';
-
-function getToken() {
-  const token = Cookies.get('AG3_JWT');
-  return token;
-}
-
-function removeToken() {
-  Cookies.remove('AG3_JWT');
-}
-
-function getMyInfo() {
-  return new Promise((resolve, reject) => {
-    const token = getToken();
-
-    if (token && !state.myInfo) {
-      axiosInstance('/users', { withCredentials: true })
-        .then(res => {
-          const myInfo = res?.data?.data;
-          if (myInfo) setState('myInfo', myInfo);
-          resolve();
-        })
-        .then(err => reject(err));
-    } else {
-      resolve();
-    }
-  });
-}
+import { getToken, setState, setToken, state } from './store';
 
 function restructingMyInfo(info) {
   const { _id, nickname, position, stacks, imageURL, sido, sigungu } = info;
@@ -47,4 +20,38 @@ function restructingMyInfo(info) {
   return myInfo;
 }
 
-export { getToken, getMyInfo, restructingMyInfo, removeToken };
+async function requestAutoLogin() {
+  try {
+    const token = getToken();
+
+    if (token && !state.myInfo) {
+      const {
+        data: { data: userInfo },
+      } = await axiosInstance('/users', { withCredentials: true });
+
+      if (userInfo) setState('myInfo', userInfo);
+    }
+  } catch (err) {
+    new Toast({
+      content: '유저정보를 불러오는데 실패하였습니다.',
+      type: 'fail',
+    });
+  }
+}
+
+async function requestLogin(user) {
+  try {
+    const res = await axiosInstance.post('/users', user, {
+      withCredentials: true,
+    });
+    const { AG3_JWT } = res.data.data;
+    const myInfo = restructingMyInfo(res.data.data);
+
+    setState('myInfo', myInfo);
+    setToken(AG3_JWT);
+  } catch (error) {
+    throw new Error(error);
+  }
+}
+
+export { restructingMyInfo, requestAutoLogin, requestLogin };
