+++ frontend/src/utils/store.js
@@ -1,3 +1,5 @@
+import Cookies from 'js-cookie';
+
 const GLOBAL_STATE_KEY = 'pro-pro-state';
 
 const savedStateJson = localStorage.getItem(GLOBAL_STATE_KEY);
@@ -22,4 +24,17 @@ function reset() {
   localStorage.removeItem(GLOBAL_STATE_KEY);
 }
 
-export { state, setState, removeState, reset };
+function getToken() {
+  const token = Cookies.get('AG3_JWT');
+  return token;
+}
+
+function setToken(token) {
+  Cookies.set('AG3_JWT', token);
+}
+
+function removeToken() {
+  Cookies.remove('AG3_JWT');
+}
+
+export { state, setState, removeState, reset, getToken, setToken, removeToken };
